refactor(rubocop): extract measure dataset filters
Issue: No ticket/issue
